src/test/java/com/devcircus/java/algorithms/sort/gnome/GnomeTest.java
package com.devcircus.java.algorithms.sort.gnome;

import com.devcircus.java.algorithms.sort.BaseSortTest;
import org.junit.Test;
import static org.junit.Assert.*;
import org.junit.Before;

/**
 *
 * @author Adrian Novegil <adrian.novegil@gmail.com>
 */
public class GnomeTest extends BaseSortTest{

    private Gnome sortter;
    
    @Before
    public void setUp() {
        sortter = new Gnome();
    }
    
    @Test
    public void testSortWithIntegers() {
        assertArrayEquals(sortter.sort(getIntegerArrayToSort()), getSortedIntegerArray());
    }

    @Test
    public void testSortWithStrings() {
        assertArrayEquals(sortter.sort(getStringArrayToSort()), getSortedStringArray());
    }
}
